+++ tests/unit/test_localized_covariance.py
